website/requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.6
django-cors-headers==4.3.1
djangorestframework==3.15.1
idna==3.7
requests==2.32.2
sqlparse==0.5.0
urllib3==2.2.1
certifi==2024.2.2
joblib==1.4.2
numpy==1.26.4
pandas==2.2.2
python-dateutil==2.9.0.post0
pytz==2024.1
scikit-learn==1.5.0
scipy==1.13.0
six==1.16.0
threadpoolctl==3.5.0
tzdata==2024.1
xgboost==2.0.3
